Move precompiled script plugin tasks into `tasks` subpackage
