+++ subprojects/plugins/src/main/kotlin/org/gradle/kotlin/dsl/plugins/precompiled/tasks/ExtractPrecompiledScriptPluginPlugins.kt
@@ -14,26 +14,47 @@
  * limitations under the License.
  */
 
-package org.gradle.kotlin.dsl.plugins.precompiled
+package org.gradle.kotlin.dsl.plugins.precompiled.tasks
 
-import org.gradle.api.file.Directory
-import org.gradle.api.provider.Provider
+import org.gradle.api.DefaultTask
+import org.gradle.api.Project
 import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.InputFiles
+import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskAction
 
+import org.gradle.kotlin.dsl.plugins.precompiled.ScriptPlugin
 
+import java.io.File
+
+
+/**
+ * Extracts the `plugins` block of each precompiled [Project] script plugin
+ * and writes it to a file with the same name under [outputDir].
+ */
 @CacheableTask
-open class CompilePrecompiledScriptPluginPlugins : ClassPathSensitiveTask() {
+open class ExtractPrecompiledScriptPluginPlugins : DefaultTask() {
 
     @get:OutputDirectory
     var outputDir = project.objects.directoryProperty()
 
+    @get:Internal
+    internal
+    lateinit var plugins: List<ScriptPlugin>
+
+    @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
+    @Suppress("unused")
+    internal
+    val scriptFiles: Set<File>
+        get() = scriptPluginFilesOf(plugins)
+
     @TaskAction
-    fun compile() {
+    fun extract() {
         outputDir.withOutputDirectory {
         }
     }
-
-    fun sourceDir(dir: Provider<Directory>) = Unit
 }
